Extract getProjectDependencies function
