+++ src/main/kotlin/com/vanniktech/dependency/graph/generator/extensions.kt
@@ -2,6 +2,7 @@ package com.vanniktech.dependency.graph.generator
 
 import org.gradle.api.Project
 import org.gradle.api.artifacts.ProjectDependency
+import com.vanniktech.dependency.graph.generator.ProjectDependencyGraphGenerator.Connection
 
 private val whitespaceRegex = Regex("\\s")
 
@@ -34,3 +35,26 @@ fun Project.isAndroidProject() = listOf("com.android.library", "com.android.appl
 fun Project.isJsProject() = plugins.hasPlugin("kotlin2js")
 
 fun Project.isCommonsProject() = plugins.hasPlugin("org.jetbrains.kotlin.platform.common")
+
+internal fun Project.getProjectDependencies(
+  projectGenerator: DependencyGraphGeneratorExtension.ProjectGenerator
+): Pair<Set<Project>, List<Connection>> {
+  val projects = mutableSetOf<Project>()
+  val dependencies = mutableListOf<Connection>()
+  fun addProject(project: Project) {
+    if (projectGenerator.includeProject(project) && projects.add(project)) {
+      project.configurations
+        .flatMap { configuration ->
+          configuration.dependencies
+            .withType(ProjectDependency::class.java)
+            .map { Connection(project, it.dependencyProject, configuration.name.endsWith("implementation", true)) }
+        }
+        .forEach {
+          dependencies.add(it)
+          addProject(it.to)
+        }
+    }
+  }
+  project.allprojects.filter { it.isDependingOnOtherProject() }.forEach { addProject(it) }
+  return projects to dependencies
+}
